feedpost component created
